Move display_rdf to RdfWrapper
